Adding permission and some changes

+++ app/Http/Controllers/User/UserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\User;
 
 use App\Http\Controllers\Controller;
+use App\Models\Modul;
 use App\Models\Package;
 use App\Models\User;
 use Illuminate\Http\Request;
@@ -16,69 +17,81 @@ class UserController extends Controller
 {
     public function index(Request $request)
     {
-        if (Auth::user()->email === 'admin@admin.com') {
-            $users = User::latest()->get();
-        } else {
+        $this->authorize('viewAny', User::class);
+
+        $users = User::latest()->get();
+
+        if (Auth::user()->email != 'admin@admin.com') {
             $users = User::where('email', '!=', 'admin@admin.com')->latest()->get();
         }
 
         if ($request->ajax()) {
             return DataTables::of($users)
                 ->addIndexColumn()
-                ->addColumn('package', function ($row) {
-                    if ($row->package_id === null) {
-                        return "--";
-                    } else {
-                        return $row->package->name;
+                ->addColumn('modules', function ($row) {
+                    $list = '<ol>';
+                    foreach ($row->modules as $modul) {
+                        $list = $list . '<li>' . $modul->code . ' - ' . $modul->name . '</li>';
                     }
+                    $list = $list . '<ol>';
+
+                    return $list;
                 })
-                ->addColumn('actions', function ($row) {
-                    $btn = '<a class="btn btn-info btn-sm" data-id="' . $row->id . '" title="Edit" id="btn-edit"><i class="ti ti-edit"></i></a>';
-                    $btn = $btn . ' <a class="btn btn-success btn-sm" data-id="' . $row->id . '" title="Detail" id="btn-info"><i class="ti ti-info-circle"></i></a>';
-                    $btn = $btn . ' <a class="btn btn-danger btn-sm" data-id="' . $row->id . '" title="Delete" id="btn-delete"><i class="ti ti-trash"></i></a>';
-                    return $btn;
+                ->addColumn('actions', function ($row) use ($request) {
+                    $btn_edit   = '';
+                    $btn_info   = '';
+                    $btn_delete = '';
+                    if ($request->user()->can('update', User::class)) {
+                        $btn_edit = '<a class="btn btn-info btn-sm" data-id="' . $row->id . '" title="Edit" id="btn-edit"><i class="ti ti-edit"></i></a>';
+                    }
+                    if ($request->user()->can('view', User::class)) {
+                        $btn_info = ' <a class="btn btn-success btn-sm" data-id="' . $row->id . '" title="Detail" id="btn-info"><i class="ti ti-info-circle"></i></a>';
+                    }
+                    if ($request->user()->can('delete', User::class)) {
+                        $btn_delete = ' <a class="btn btn-danger btn-sm" data-id="' . $row->id . '" title="Delete" id="btn-delete"><i class="ti ti-trash"></i></a>';
+                    }
+                    return $btn_edit . $btn_info . $btn_delete;
                 })
-                ->rawColumns(['packages', 'actions'])
+                ->rawColumns(['modules', 'actions'])
                 ->make(true);
         }
 
         return view('settings.user.index', [
             'title'    => 'Users',
             'name'     => Auth::user()->name,
-            'packages' => Package::all()
+            'modules'  => Modul::all()->sortBy('sequence')
         ]);
-    }
-
-    public function getPackages()
-    {
-        $packages = Package::all();
-        return response()->json($packages);
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', User::class);
+
         $validator = Validator::make($request->all(), [
             'name'       => 'required',
             'email'      => 'required|email:rfc,dns|unique:users',
             'username'   => 'required|unique:users',
             'password'   => 'required|confirmed',
-            'package_id' => 'required'
+            'modul_id'   => 'sometimes',
         ], [
-            'package_id.required' => 'Please choose the package'
+            'modul_id.required' => 'Please choose the modul'
         ]);
 
         if ($validator->fails()) {
             return response()->json($validator->errors(), 422);
         }
 
-        $package       = Package::find($request->package_id);
         $user           = new User();
-        $user->name     = ucwords($request->name);
+        $user->name     = $request->name;
         $user->email    = $request->email;
         $user->username = $request->username;
         $user->password = Hash::make($request->password);
-        $user->package()->associate($package);
         $user->save();
+
+        if ($request->modul_id) {
+            $modules = Modul::find($request->modul_id);
+            $user->modules()->sync($modules);
+        }
 
 
         return response()->json([
@@ -89,57 +102,50 @@ class UserController extends Controller
 
     public function show(User $user)
     {
-        if ($user->package_id === null) {
-            $package = null;
-        } else {
-            $package = $user->package_id;
-        }
+        $this->authorize('view', User::class);
 
         return response()->json([
             'success'  => true,
             'data'     => $user,
-            'package'  => $package,
+            'modules'  => $user->modules->pluck('id'),
             'info'     => $this->info($user)
         ]);
     }
 
     public function info(User $user)
     {
-        if ($user->package_id === null) {
-            $list = null;
-        } else {
-            $list = '<ul>';
-            foreach ($user->package->moduls as $modul) {
-                $list = $list . '<li data-jstree=\'{"opened":true, "icon":"ti ti-folder"}\'>' . $modul->code . ' - ' . $modul->name;
+        $list = '<ul>';
+        foreach ($user->modules as $modul) {
+            $list = $list . '<li data-jstree=\'{"opened":true, "icon":"ti ti-folder"}\'>' . $modul->code . ' - ' . $modul->name;
+            $list = $list . '<ul>';
+            foreach ($modul->menus as $menu) {
+                $list = $list . '<li data-jstree=\'{"opened":true, "icon":"ti ti-' . $menu->icon . '"}\'>' . $menu->code . ' - ' . $menu->name;
                 $list = $list . '<ul>';
-                foreach ($modul->menus as $menu) {
-                    $list = $list . '<li data-jstree=\'{"opened":true, "icon":"ti ti-' . $menu->icon . '"}\'>' . $menu->code . ' - ' . $menu->name;
-                    $list = $list . '<ul>';
-                    foreach ($menu->permissions as $permission) {
-                        $list = $list . '<li data-jstree=\'{"opened":true, "icon":"ti ti-fingerprint"}\'>' . $permission->name . '</li>';
-                    }
-                    $list = $list . '</ul>';
+                foreach ($menu->permissions as $permission) {
+                    $list = $list . '<li data-jstree=\'{"opened":true, "icon":"ti ti-fingerprint"}\'>' . $permission->name . '</li>';
                 }
                 $list = $list . '</ul>';
-                $list = $list . '</li>';
             }
             $list = $list . '</ul>';
+            $list = $list . '</li>';
         }
-
+        $list = $list . '</ul>';
 
         return $list;
     }
 
     public function update(Request $request, User $user)
     {
+        $this->authorize('update', User::class);
+
         $validator = Validator::make($request->all(), [
             'name'       => 'required',
             'email'      => ['required', 'email:rfc,dns', Rule::unique('users')->ignore($user->id)],
             'username'   => ['required', Rule::unique('users')->ignore($user->id)],
             'password'   => 'sometimes|confirmed',
-            'package_id' => 'required'
+            'modul_id'   => 'sometimes'
         ], [
-            'package_id.required' => 'Please choose the package'
+            'modul_id.required' => 'Please choose the modul'
         ]);
 
         if ($validator->fails()) {
@@ -150,13 +156,16 @@ class UserController extends Controller
             $user->password = Hash::make($request->password);
         }
 
-        $package        = Package::find($request->package_id);
-        $user->name     = ucwords($request->name);
+        $user->name     = $request->name;
         $user->email    = $request->email;
         $user->username = $request->username;
         $user->password = $user->password;
-        $user->package()->associate($package);
         $user->save();
+
+        if ($request->modul_id) {
+            $modules = Modul::find($request->modul_id);
+            $user->modules()->sync($modules);
+        }
 
         return response()->json([
             'success' => true,
@@ -166,8 +175,9 @@ class UserController extends Controller
 
     public function destroy(User $user)
     {
-        $user->delete();
+        $this->authorize('delete', User::class);
 
+        $user->delete();
         return response()->json([
             'success' => true,
             'data'    => $user
